requirements.txt
google-generativeai==0.5.4
Pillow==9.3.0
streamlit==1.34.0
fpdf2
python-dotenv